+++ test/groovy/org/ods/orchestration/usecase/JiraUseCaseSpec.groovy
@@ -2,6 +2,7 @@ package org.ods.orchestration.usecase
 
 
 import org.ods.orchestration.service.JiraService
+import org.ods.orchestration.util.TestResults
 import org.ods.util.IPipelineSteps
 import org.ods.util.ILogger
 import org.ods.util.Logger
@@ -822,8 +823,9 @@ class JiraUseCaseSpec extends SpecHelper {
         usecase.updateJiraReleaseStatusBuildNumber()
 
         then:
-        1 * jira.updateReleaseStatusIssue(_, 'someChangeId', [
-            buildNumber: "1.0-0815"
+        1 * jira.updateBuildNumber(_, 'someChangeId', [
+            buildNumber: "1.0-0815",
+            env: "D",
         ])
     }
 
@@ -833,6 +835,10 @@ class JiraUseCaseSpec extends SpecHelper {
         project.buildParams.version = "1.0"
         steps.env.BUILD_NUMBER = "0815"
         steps.env.RUN_DISPLAY_URL = "http://jenkins"
+        def TestResults testResults = new TestResults();
+        testResults.setSucceeded(1)
+        testResults.setFailed(1)
+        project.setAggregatedTestResults(testResults)
 
         def error = new RuntimeException("Oh no!")
 
@@ -841,6 +847,14 @@ class JiraUseCaseSpec extends SpecHelper {
 
         then:
         1 * jira.updateReleaseStatusIssue(_, 'someChangeId', [
+            userEmail: null,
+            testResults: [
+                skipped: testResults.skipped,
+                succeeded: testResults.succeeded,
+                failed: testResults.failed,
+                error: testResults.error,
+                missing: testResults.missing,
+            ],
             status: "Failed",
             env: 'D',
         ])
@@ -855,12 +869,23 @@ class JiraUseCaseSpec extends SpecHelper {
         project.data.buildParams.changeId = "someChangeId"
         project.buildParams.version = "1.0"
         steps.env.BUILD_NUMBER = "0815"
+        def TestResults testResults = new TestResults();
+        testResults.setSucceeded(1)
+        project.setAggregatedTestResults(testResults)
 
         when:
         usecase.updateJiraReleaseStatusResult("", false)
 
         then:
         1 * jira.updateReleaseStatusIssue(_, 'someChangeId', [
+            userEmail: null,
+            testResults: [
+                skipped: testResults.skipped,
+                succeeded: testResults.succeeded,
+                failed: testResults.failed,
+                error: testResults.error,
+                missing: testResults.missing,
+            ],
             status: "Successful",
             env: 'D',
         ])